requirements.txt
time
selenium
mysql.connector
getpass
